move truncateMiddle into Truncate
