+++ src/utilities/strings.js
@@ -50,32 +50,6 @@ export const textIncludesOnlyEmoji = text => {
   return !withSpacesRemoved
 }
 
-// Source
-// https://github.com/kahwee/truncate-middle
-export const truncateMiddle = (word, startLen, endLen, ellip) => {
-  if (!isWordString(word)) {
-    return ''
-  }
-  const wordLen = word.length
-  // Setting default values
-  const frontLen = ~~startLen // will cast to integer
-  const backLen = ~~endLen
-  const truncateStr = ellip !== undefined ? ellip : '…'
-
-  if (
-    (frontLen === 0 && backLen === 0) ||
-    frontLen >= wordLen ||
-    backLen >= wordLen ||
-    frontLen + backLen >= wordLen
-  ) {
-    return word
-  } else if (backLen === 0) {
-    return word.slice(0, frontLen) + truncateStr
-  } else {
-    return word.slice(0, frontLen) + truncateStr + word.slice(wordLen - backLen)
-  }
-}
-
 export const stripUrlPrefix = url => {
   if (!isString(url)) return url
   return url.replace(
